Move dialogue, objects, speccurses into src
